Added CourseDegree delRecord()
Need to review how the coursedegree class should be treated in the overall DDL. It currently serves as nothing but as the connection between the course table and the degree table, which is then queryable from the enrollment module. I will be developing this initially using a static method to get courses. Should the group decide on another design/implementation for this, I'll modify it accordingly.